refactor(openpanel): resolve service-layer circular dependency
