+++ apps/connect/src/services/openpanel/index.ts
@@ -1,3 +1,4 @@
+export * from "./buffer.js";
 export * from "./client.js";
 export * from "./events.js";
 export * from "./factory.js";
