+++ apps/connect/src/services/openpanel/types.ts
@@ -5,6 +5,12 @@
 export type OpenPanelConfig = {
   clientId: string;
   apiUrl?: string;
+  /**
+   * Reserved for future UI-event tracking (button clicks, navigation, etc.).
+   * The flag is wired into `connectConfig` but no call sites gate on it yet —
+   * those will be added in a follow-up card.  It is intentionally dormant and
+   * should not be mistaken for a bug.
+   */
   trackUiEvents: boolean;
   trackOperations: boolean;
 };
